refactor(recent-food): extract touch functions and centralize dependencies
